Converting WaypointDownloadsPage to TS
