--- src/pages/_proxied-dot-io/waypoint/downloads/index.jsx
@@ -1,55 +1,0 @@
-import WaypointIoLayout from 'layouts/_proxied-dot-io/waypoint'
-import productData from 'data/waypoint'
-import ProductDownloadsPage from '@hashicorp/react-product-downloads-page'
-import { generateStaticProps } from 'lib/fetch-release-data'
-import styles from './style.module.css'
-
-function DownloadsPage({ releases, latestVersion }) {
-  return (
-    <ProductDownloadsPage
-      releases={releases}
-      packageManagers={productData.packageManagers}
-      productName={productData.name}
-      productId={productData.slug}
-      latestVersion={latestVersion}
-      getStartedDescription="Follow step-by-step tutorials on AWS, Azure, GCP, and localhost."
-      getStartedLinks={[
-        {
-          label: 'Deploy to Docker',
-          href:
-            'https://learn.hashicorp.com/collections/waypoint/get-started-docker',
-        },
-        {
-          label: 'Deploy to Kubernetes',
-          href:
-            'https://learn.hashicorp.com/collections/waypoint/get-started-kubernetes',
-        },
-        {
-          label: 'Deploy to AWS',
-          href: 'https://learn.hashicorp.com/collections/waypoint/deploy-aws',
-        },
-        {
-          label: 'View all Waypoint tutorials',
-          href: 'https://learn.hashicorp.com/waypoint',
-        },
-      ]}
-      logo={
-        <img
-          className={styles.logo}
-          alt="Waypoint"
-          src={require('./img/waypoint-logo.svg')}
-        />
-      }
-      product="waypoint"
-      tutorialLink={{
-        href: 'https://learn.hashicorp.com/waypoint',
-        label: 'View Tutorials at HashiCorp Learn',
-      }}
-    />
-  )
-}
-
-export const getStaticProps = () => generateStaticProps('waypoint')
-
-DownloadsPage.layout = WaypointIoLayout
-export default DownloadsPage
